job_signal_ai/requirements.txt
streamlit>=1.28.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
serpapi>=0.1.0
openai>=1.0.0
